fix: 肄붾뱶由щ럭 諛섏쁺
